atualizando

+++ login/microservicologin/src/main/java/com/br/unisales/microservicologin/service/UsuarioService.java
@@ -15,7 +15,6 @@ import com.br.unisales.microservicologin.table.Usuario;
 /**
  * @apiNote Classe responsável por responder as solicitações do controller em relação ao usuário
  * 
- * @author Vito Rodrigues Franzosi
  * @Data Criação 24.04.2024
  */
 @Service
@@ -29,7 +28,6 @@ public class UsuarioService {
      * @param String senha
      * @return UsuarioModel
      * 
-     * @author Vito Rodrigues Franzosi
      * @Data Criação 24.04.2024
      */
     public UsuarioModel login(String email, String senha) {
@@ -47,7 +45,6 @@ public class UsuarioService {
      * @apiNote Método responsável por listar todos os usuários cadastrados
      * @return List<UsuarioModel>
      * 
-     * @author Vito Rodrigues Franzosi
      * @Data Criação 24.04.2024
      */
     public List<UsuarioModel> listar() {
@@ -62,7 +59,6 @@ public class UsuarioService {
      * @apiNote Método responsável por listar todos os usuários cadastrados
      * @return List<UsuarioModel>
      * 
-     * @author Vito Rodrigues Franzosi
      * @Data Criação 24.04.2024
      */
     public List<UsuarioModel> listarPorParametros(String nome, String sexo, String email, String grupo) {
@@ -100,7 +96,6 @@ public class UsuarioService {
      * @apiNote Método responsável por buscar o usuário pelo seu id
      * @return UsuarioModel
      * 
-     * @author Vito Rodrigues Franzosi
      * @Data Criação 24.04.2024
      */
     public UsuarioModel buscarPorId(Integer id) {
@@ -120,7 +115,6 @@ public class UsuarioService {
      * @param String grupo
      * @return UsuarioModel
      * 
-     * @author Vito Rodrigues Franzosi
      * @Data Criação 28.04.2024
      */
     public UsuarioModel salvar(Integer id, String nome, String sexo, String email, String senha, String grupo) {
@@ -133,8 +127,7 @@ public class UsuarioService {
         usuario.setEmail(email);
         usuario.setGrupo(grupo);
         usuario.setNome(nome);
-        if((senha.trim()).length()>0)
-            usuario.setSenha(senha);
+        if(!(senha.trim()).isEmpty()) usuario.setSenha(senha);
         usuario.setSexo(sexo);
         usuario = this.repo.save(usuario);
         return this.converterUsuarioToModel(usuario);
@@ -144,13 +137,11 @@ public class UsuarioService {
      * @apiNote Método responsável por excluir um usuário do banco de dados
      * @param Integer id
      * 
-     * @author Vito Rodrigues Franzosi
      * @Data Criação 28.04.2024
      */
     public void excluir(Integer id) {
         Optional<Usuario> usuarioOptional = this.repo.findById(id);
-        if(usuarioOptional.isPresent())
-            this.repo.delete(usuarioOptional.get());
+        if(usuarioOptional.isPresent()) this.repo.delete(usuarioOptional.get());
     }
 
     /**
@@ -158,7 +149,6 @@ public class UsuarioService {
      * @param Usuario usuario
      * @return UsuarioModel
      * 
-     * @author Vito Rodrigues Franzosi
      * @Data Criação 24.04.2024
      */
     private UsuarioModel converterUsuarioToModel(Usuario usuario) {
